perf_monitor/requirements.txt
mysql-connector-python
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
python-dotenv
